back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = ""
authors = [{ name = "Reza Rahemtola", email = "reza.rahemtola@epitech.eu" }]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "web3 (>=7.12.0,<8.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "eth-account (>=0.13.7,<0.14.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
